Fix bug with email notification
If checkbox is checked the web browser will send

~~~
{
    ...,
    "send_email_field": "on",
    ...
}
~~~

but otherwise it will not include "send_email_field"
on the request what make hard to know if we should
or not send the email.

This pull request change the question
to make easy to know when the email should be send.